symlink sbosnap and supplement local repo check (#15)
Instead of simply checking for SLACKBUILDS.TXT, check the local repository for similarity to upstream as well; provide more informative prompts depending on the repo contents and fetch method.

The distinction between sbosnap and sbocheck was always based on the presence or absence of SLACKBUILDS.TXT. Now that errors resulting from the absence of this file are preceded by prompts and a different checking process is in place, it is no longer necessary to keep both of them. sbosnap and sbosnap.1.gz will be installed as symlinks for compatibility.